[ADD] Validators of pieces and view properly drawn.

+++ modelo/Abscisa.java
@@ -2,12 +2,12 @@ package ajedrez;
 
 public enum Abscisa {
     a(0),b(1),c(2),d(3),e(4),f(5),g(6),h(7);
-    private int value;
-    private Abscisa(int _value){
+    private byte value;
+    private Abscisa(byte _value){
         this.value = _value;
     }
 
-    public int getValue(){
+    public byte getValue(){
         return value;
     }
 }
